move remove_overlaps to outline transforms
