+++ torchfont/transforms/curves.py
@@ -1,4 +1,4 @@
-"""Bezier curve format conversion, segment merging, and outline simplification."""
+"""Bezier curve format conversion and segment merging."""
 
 import torch
 from torch import Tensor
@@ -108,31 +108,6 @@ def merge_curves(types: Tensor, coords: Tensor) -> tuple[Tensor, Tensor]:
     types = types.cpu().contiguous()
     coords = coords.cpu().contiguous()
     out_types, out_coords = _torchfont.merge_curves(
-        types.numpy(), coords.reshape(-1).numpy()
-    )
-    return (
-        torch.tensor(out_types, dtype=torch.long),
-        torch.tensor(out_coords, dtype=torch.float32).view(-1, 6),
-    )
-
-
-def remove_overlaps(types: Tensor, coords: Tensor) -> tuple[Tensor, Tensor]:
-    """Merge overlapping subpaths using Skia PathOps winding simplification.
-
-    Args:
-        types: 1-D ``torch.int64`` tensor of path element types.
-        coords: 2-D ``torch.float32`` tensor of shape ``(N, 6)``.
-
-    Returns:
-        A new variable-length outline tuple ``(types, coords)`` with overlapping
-        subpath edges removed when Skia PathOps can resolve the outline. If
-        PathOps cannot simplify an otherwise valid outline, the original outline
-        is returned unchanged.
-
-    """
-    types = types.cpu().contiguous()
-    coords = coords.cpu().contiguous()
-    out_types, out_coords = _torchfont.remove_overlaps(
         types.numpy(), coords.reshape(-1).numpy()
     )
     return (
